Initial version by students
This initial version was designed and devolped by the following
students:
- Max Hamulyak
- Bjorn Broer
- Stef Chappin
- Bert van Montfort
- Guy
Spronck

They created it as part of their Cross-Platform Mobile Development (code
MIT4) class at Zuyd University for Zuyd University.

The goal was to create an app in which users can learn programming by
playing a game and that's what we have created.

Here is a video showing
what the app looks like, and what it does:
[ProgramADroid](https://drive.google.com/open?id=0B1AsMmKz2KeOZ0RUYVJnOUJtRzg)